trip-itinerary-planner/pyproject.toml
[project]
name = "itinerary-planner"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-langgraph>=0.0.32",
    "grpcio>=1.75.1",
    "langgraph>=0.3.25",
    "langchain>=0.3.14",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "nacos-sdk-python>=3.0.4",
    # Pin a2a-sdk: nacos-sdk-python's dependency declaration is unbounded so
    # uv would otherwise resolve to a2a-sdk 1.0.0 whose package layout drops
    # `a2a.types.AgentCard`, breaking nacos-sdk-python's import chain.
    # Mirror the version chat-service is already running on.
    "a2a-sdk==0.3.22",
    "openinference-instrumentation-langchain>=0.1.58",
    "opentelemetry-distro[otlp]>=0.58b0",
    "pydantic-settings>=2.11.0",
    "tripsphere>=0.1.0",
    "uvicorn>=0.40.0",
    "fastapi>=0.135.1",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "types-grpcio>=1.0.0.20251009",
    "types-protobuf>=6.32.1.20251210",
]

[tool.uv.sources]
tripsphere = { workspace = true }

[tool.uv.workspace]
members = ["libs/tripsphere"]

[tool.ruff]
target-version = "py312"
extend-exclude = ["libs/tripsphere"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q"

[tool.pyright]
typeCheckingMode = "strict"
include = ["src"]
